pyproject.toml
[project]
name = "hybrid-job-matching-mock-interview-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.6.1",
    "flask>=3.1.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.27.0",
    "google-generativeai>=0.8.5",
    "mysqlclient>=2.2.7",
    "pdfplumber>=0.11.7",
    "pymysql>=1.1.1",
    "pypdf>=5.8.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.4",
    "six>=1.17.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "mypy>=1.17.0",
    "pytest>=8.4.1",
    "pytest-flask>=1.3.0",
]
